requirements.txt
# CMS振动分析报告系统依赖包
# 版本: 2.1.0
# 更新日期: 2025-01-16
# 精确版本匹配当前环境，确保完全兼容
# 
# 安装说明:
# 1. 建议使用Python 3.12环境
# 2. 推荐使用虚拟环境: python -m venv venv && source venv/bin/activate
# 3. 安装命令: pip install -r requirements.txt
# 4. 如遇安装问题，请参考README.md中的故障排除部分
#
# 系统要求:
# - Python: 3.8+ (推荐3.12)
# - 内存: 16GB+ (模型加载)
# - 存储: 20GB+ (模型文件)
# - CUDA: 12.6+ (GPU加速，可选)

# ==========================================
# 核心Web框架和界面
# ==========================================
streamlit==1.48.0          # 主要Web界面框架
gradio==5.9.1              # 备用界面框架

# ==========================================
# 数据处理和科学计算
# ==========================================
numpy==2.2.6               # 数值计算基础库
pandas==2.3.1              # 数据处理和分析
scipy==1.16.1              # 科学计算库（信号处理）

# ==========================================
# 机器学习和深度学习框架
# ==========================================
torch==2.7.1               # PyTorch深度学习框架
transformers==4.55.0       # HuggingFace Transformers库
sentence-transformers==5.1.0  # 句子嵌入模型
accelerate==1.10.0         # 模型加速库
tokenizers==0.21.4         # 快速分词器
huggingface-hub==0.34.4    # HuggingFace模型中心
safetensors==0.6.2         # 安全张量存储格式

# ==========================================
# 向量数据库和检索
# ==========================================
chromadb==1.0.16           # 向量数据库（RAG核心）

# ==========================================
# 数据可视化和图表生成
# ==========================================
matplotlib==3.10.5         # 基础绘图库
seaborn==0.13.2            # 统计图表库
plotly==6.2.0              # 交互式图表库

# ==========================================
# 报告生成和文档处理
# ==========================================
reportlab==4.4.3           # PDF报告生成
python-docx==1.2.0         # Word文档生成
Jinja2==3.1.6              # 模板引擎
weasyprint==66.0           # HTML到PDF转换
MarkupSafe==3.0.2          # 模板安全处理

# ==========================================
# 图像处理
# ==========================================
Pillow==11.3.0             # 图像处理库

# ==========================================
# 日志记录和调试
# ==========================================
loguru==0.7.3              # 高级日志记录库

# ==========================================
# 机器学习和API服务
# ==========================================
scikit-learn==1.7.1        # 机器学习算法库
fastapi==0.116.1           # 现代API框架
uvicorn==0.35.0            # ASGI服务器

# ==========================================
# 数据库和ORM
# ==========================================
SQLAlchemy==2.0.42         # SQL工具包和ORM

# ==========================================
# HTTP客户端和网络
# ==========================================
httpx==0.28.1              # 现代HTTP客户端
requests==2.31.0           # 传统HTTP客户端

# ==========================================
# 配置管理和文件格式
# ==========================================
PyYAML==6.0.2              # YAML配置文件解析
toml==0.10.2               # TOML配置文件解析

# ==========================================
# 工具库和实用程序
# ==========================================
tqdm==4.66.2               # 进度条显示
typing_extensions==4.14.1  # 类型注解扩展
click==8.2.1               # 命令行界面框架
rich==14.1.0               # 富文本终端输出
filelock==3.18.0           # 文件锁定
fsspec==2025.7.0           # 文件系统规范
regex==2025.7.33           # 正则表达式增强
psutil==7.0.0              # 系统和进程工具
packaging==24.2            # 包装工具
setuptools==79.0.1         # 安装工具
wheel==0.43.0              # 包分发格式

# ==========================================
# 安全和认证
# ==========================================
cryptography==41.0.7       # 加密和安全库

# ==========================================
# AI和语言模型集成
# ==========================================
openai==1.90.0             # OpenAI API客户端
langchain==0.3.27          # LangChain框架
langchain-community==0.3.27 # LangChain社区组件
langchain-core==0.3.74     # LangChain核心组件
langchain-text-splitters==0.3.9 # 文本分割器

# ==========================================
# 开发和测试工具（可选）
# ==========================================
# pytest==7.4.3            # 单元测试框架
# black==23.12.1           # 代码格式化
# flake8==6.1.0            # 代码检查
# mypy==1.8.0              # 类型检查
# pre-commit==3.6.0        # Git钩子管理

# ==========================================
# 安装验证命令
# ==========================================
# 安装完成后，运行以下命令验证:
# python -c "import streamlit, torch, transformers, chromadb; print('✅ 核心依赖安装成功')"
# python -c "import matplotlib, reportlab, docx; print('✅ 报告生成依赖安装成功')"
# python -c "import torch; print(f'CUDA可用: {torch.cuda.is_available()}')"

# ==========================================
# 故障排除提示
# ==========================================
# 如果遇到安装问题:
# 1. weasyprint安装失败: sudo apt-get install libpango-1.0-0 libharfbuzz0b libpangoft2-1.0-0
# 2. torch版本冲突: pip install torch --upgrade --force-reinstall
# 3. 内存不足: 考虑使用CPU版本或启用模型量化
# 4. 权限问题: 使用 --user 参数或虚拟环境
# 5. 网络问题: 使用国内镜像源 -i https://pypi.tuna.tsinghua.edu.cn/simple/